requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-easy-pjax==1.2.0
django-material==0.4.1
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
mysqlclient==1.3.6
Pillow==2.9.0
psycopg2==2.6.1
static3==0.6.1
Unipath==1.1
wheel==0.24.0
